backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.6.0
python-dotenv>=1.0.1
pydantic>=2.8.0
email-validator>=2.2.0
python-multipart
onnxruntime==1.17.0
Pillow
numpy
httpx
pytest
onnx
websockets
slowapi==0.1.9
pydantic-settings>=2.2.1
